Remove unnecessary casts

+++ edgraph/server.go
@@ -2152,7 +2152,7 @@ func verifyUniqueWithinMutation(qc *queryContext) error {
 	for i := range qc.uniqueVars {
 		gmuIndex, rdfIndex := decodeIndex(i)
 		// handles cases where the mutation was pruned in updateMutations
-		if gmuIndex >= uint32(len(qc.gmuList)) || qc.gmuList[gmuIndex] == nil || uint32(rdfIndex) >= uint32(len(qc.gmuList[gmuIndex].Set)) {
+		if gmuIndex >= uint32(len(qc.gmuList)) || qc.gmuList[gmuIndex] == nil || rdfIndex >= uint32(len(qc.gmuList[gmuIndex].Set)) {
 			continue
 		}
 		pred1 := qc.gmuList[gmuIndex].Set[rdfIndex]
@@ -2163,7 +2163,7 @@ func verifyUniqueWithinMutation(qc *queryContext) error {
 			}
 			gmuIndex2, rdfIndex2 := decodeIndex(j)
 			// bounds check for the second predicate, which could also have been pruned
-			if gmuIndex2 >= uint32(len(qc.gmuList)) || qc.gmuList[gmuIndex2] == nil || uint32(rdfIndex2) >= uint32(len(qc.gmuList[gmuIndex2].Set)) {
+			if gmuIndex2 >= uint32(len(qc.gmuList)) || qc.gmuList[gmuIndex2] == nil || rdfIndex2 >= uint32(len(qc.gmuList[gmuIndex2].Set)) {
 				continue
 			}
 			pred2 := qc.gmuList[gmuIndex2].Set[rdfIndex2]
